% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_table.R
\name{eigen_table}
\alias{eigen_table}
\title{Create a tibble of eigenvalues from a list of matrices}
\usage{
eigen_table(mat_list, prefix = "lambda", mat_name = "component")
}
\arguments{
\item{mat_list}{Named list of square matrices.}

\item{prefix}{Column name prefix.}

\item{mat_name}{Name of column containing names of \code{mat_list}.}
}
\description{
Create a tibble of eigenvalues from a list of matrices
}
