% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covs.R
\name{get_covs}
\alias{get_covs}
\title{Extract estimated covariance matrices from \code{lme()} fit}
\usage{
get_covs(fit)
}
\arguments{
\item{fit}{Object returned by \code{lme()} with random effect formula with grouping `sire/dam'}
}
\description{
Extract estimated covariance matrices from \code{lme()} fit
}
